refactor: move ToolCallArguments to a separate file and optimize its implementation
