+++ crates/forge_domain/src/lib.rs
@@ -31,6 +31,7 @@ mod task;
 mod temperature;
 mod template;
 mod tool_call;
+mod tool_call_arguments;
 mod tool_call_context;
 mod tool_choice;
 mod tool_definition;
@@ -77,6 +78,7 @@ pub use task::*;
 pub use temperature::*;
 pub use template::*;
 pub use tool_call::*;
+pub use tool_call_arguments::*;
 pub use tool_call_context::*;
 pub use tool_choice::*;
 pub use tool_definition::*;
